src/api/requirements.txt
fastapi
uvicorn
pillow
transformers
torch
torchvision
timm
flash_attn
einops